Cleanup teamcity
