+++ .teamcity/IMODCollector/Project.kt
@@ -1,49 +1,17 @@
 package IMODCollector
 
-import IMODCollector.buildTypes.*
-import jetbrains.buildServer.configs.kotlin.BuildType
-
+import IMODCollector.buildTypes.Coupler_Regression_Binaries
+import IMODCollector.buildTypes.IMODCollector_X64development
+import IMODCollector.buildTypes.Ribasim_binaries
 import jetbrains.buildServer.configs.kotlin.Project
-import jetbrains.buildServer.configs.kotlin.buildSteps.script
 
 object Project : Project({
     id("IMODCollector")
     name = "iMOD_Collector"
     description = "Collect iMOD6 coupled components + coupler into a single package"
+
     buildType(IMODCollector_X64development)
     buildType(Ribasim_binaries)
     buildType(Coupler_Regression_Binaries)
 })
 
-object Ribasim_binaries : BuildType({
-    name = "Ribasim binaries"
-    description = "Download Ribasim release binaries"
-
-    artifactRules = """
-        ribasim/** => ribasim.zip!
-    """.trimIndent()
-
-    params {
-        select("RIBASIM_Platform", "",
-            options = listOf("windows", "linux"))
-        select("RIBASIM_Version", "",
-            options = listOf("v2025.6.0", "v2025.5.0", "v2025.4.0", "v2025.3.0"))
-    }
-
-    vcs {
-    }
-
-    steps {
-        script {
-            name = "Download Ribasim"
-            scriptContent = """
-                wget https://github.com/Deltares/Ribasim/releases/download/%RIBASIM_Version%/ribasim_%RIBASIM_Platform%.zip -O ribasim.zip
-                unzip  "ribasim.zip"
-            """.trimIndent()
-        }
-    }
-
-    requirements {
-        equals("teamcity.agent.jvm.os.name", "Linux")
-    }
-})
