+++ .teamcity/_Self/buildTypes/MyPy.kt
@@ -1,9 +1,9 @@
 package _Self.buildTypes
 
+import Templates.GitHubIntegrationTemplate
 import _Self.vcsRoots.ImodCoupler
 import jetbrains.buildServer.configs.kotlin.BuildType
 import jetbrains.buildServer.configs.kotlin.buildFeatures.XmlReport
-import jetbrains.buildServer.configs.kotlin.buildFeatures.commitStatusPublisher
 import jetbrains.buildServer.configs.kotlin.buildFeatures.xmlReport
 import jetbrains.buildServer.configs.kotlin.buildSteps.script
 import jetbrains.buildServer.configs.kotlin.failureConditions.BuildFailureOnMetric
@@ -12,8 +12,10 @@ import jetbrains.buildServer.configs.kotlin.failureConditions.failOnMetricChange
 object MyPy : BuildType({
     name = "MyPy"
 
+    templates(GitHubIntegrationTemplate)
+
     vcs {
-        root(_Self.vcsRoots.ImodCoupler, ". => imod_coupler")
+        root(ImodCoupler, ". => imod_coupler")
         cleanCheckout = true
     }
 
@@ -53,17 +55,6 @@ object MyPy : BuildType({
     }
 
     features {
-        commitStatusPublisher {
-            id = "BUILD_EXT_142"
-            vcsRootExtId = "${ImodCoupler.id}"
-            publisher = github {
-                githubUrl = "https://api.github.com"
-                authType = personalToken {
-                    token = "credentialsJSON:6b37af71-1f2f-4611-8856-db07965445c0"
-                }
-            }
-        }
-
         xmlReport {
             reportType = XmlReport.XmlReportType.JUNIT
             rules = "imod_coupler/*.xml"
