+++ .teamcity/_Self/buildTypes/TwineCheck.kt
@@ -1,12 +1,14 @@
 package _Self.buildTypes
 
+import Templates.GitHubIntegrationTemplate
 import _Self.vcsRoots.ImodCoupler
 import jetbrains.buildServer.configs.kotlin.BuildType
-import jetbrains.buildServer.configs.kotlin.buildFeatures.commitStatusPublisher
 import jetbrains.buildServer.configs.kotlin.buildSteps.script
 
 object TwineCheck : BuildType({
     name = "Twine"
+
+    templates(GitHubIntegrationTemplate)
 
     vcs {
         root(ImodCoupler, ". => imod_coupler")
@@ -22,19 +24,6 @@ object TwineCheck : BuildType({
                     pixi run check-package-primod
                 """.trimIndent()
             formatStderrAsError = true
-        }
-    }
-
-    features {
-        commitStatusPublisher {
-            id = "BUILD_EXT_142"
-            vcsRootExtId = "${ImodCoupler.id}"
-            publisher = github {
-                githubUrl = "https://api.github.com"
-                authType = personalToken {
-                    token = "credentialsJSON:6b37af71-1f2f-4611-8856-db07965445c0"
-                }
-            }
         }
     }
