+++ .teamcity/_Self/vcsRoots/ImodCoupler.kt
@@ -1,6 +1,5 @@
 package _Self.vcsRoots
 
-import jetbrains.buildServer.configs.kotlin.*
 import jetbrains.buildServer.configs.kotlin.vcs.GitVcsRoot
 
 object ImodCoupler : GitVcsRoot({
@@ -13,7 +12,7 @@ object ImodCoupler : GitVcsRoot({
         -:refs/heads/gh-pages
     """.trimIndent()
     useTagsAsBranches = true
-    checkoutPolicy = GitVcsRoot.AgentCheckoutPolicy.USE_MIRRORS
+    checkoutPolicy = AgentCheckoutPolicy.USE_MIRRORS
     authMethod = password {
         userName = "teamcity-deltares"
         password = "credentialsJSON:abf605ce-e382-4b10-b5de-8a7640dc58d9"
